NMS-10439 adopt for develop branch
