+++ features/enlinkd/persistence/api/src/main/java/org/opennms/netmgt/enlinkd/model/CdpLinkTopologyEntity.java
@@ -90,7 +90,7 @@ public class CdpLinkTopologyEntity {
 
     @Override
     public String toString() {
-        return "CdpLinkInfo{" +
+        return "CdpLinkTopologyEntity{" +
                 "id=" + id +
                 ", nodeId=" + nodeId +
                 ", cdpCacheIfIndex=" + cdpCacheIfIndex +
