+++ features/enlinkd/persistence/api/src/main/java/org/opennms/netmgt/enlinkd/model/LldpLinkTopologyEntity.java
@@ -26,9 +26,7 @@
  *     http://www.opennms.com/
  *******************************************************************************/
 
-package org.opennms.netmgt.model;
-
-import java.util.Objects;
+package org.opennms.netmgt.enlinkd.model;
 
 import org.opennms.core.utils.LldpUtils;
 
@@ -62,8 +60,8 @@ public class LldpLinkTopologyEntity {
         this.lldpPortIfindex = lldpPortIfindex;
     }
 
-    public LldpLinkTopologyEntity(LldpLink link) {
-        this(link.getId()
+    public static LldpLinkTopologyEntity create(LldpLink link) {
+        return new LldpLinkTopologyEntity(link.getId()
                 , link.getNode().getId()
                 , link.getLldpRemChassisId()
                 , link.getLldpRemPortId()
