+++ features/enlinkd/persistence/api/src/main/java/org/opennms/netmgt/enlinkd/persistence/api/TopologyEntityDao.java
@@ -32,6 +32,10 @@ import java.util.List;
 
 import org.opennms.netmgt.enlinkd.model.CdpLinkTopologyEntity;
 import org.opennms.netmgt.enlinkd.model.NodeTopologyEntity;
+import org.opennms.netmgt.enlinkd.model.IsIsLinkTopologyEntity;
+import org.opennms.netmgt.enlinkd.model.LldpLinkTopologyEntity;
+import org.opennms.netmgt.enlinkd.model.OspfLinkTopologyEntity;
+
 
 /**
  * Retrieves TopologyEntities from the database. TopologyEntities are views on OnmsEntities (such as OnmsNode, CdpLink, etc.)
@@ -44,4 +48,7 @@ import org.opennms.netmgt.enlinkd.model.NodeTopologyEntity;
 public interface TopologyEntityDao {
     List<NodeTopologyEntity> getNodeTopologyEntities();
     List<CdpLinkTopologyEntity> getCdpLinkTopologyEntities();
+    List<IsIsLinkTopologyEntity> getIsIsLinkTopologyEntities();
+    List<LldpLinkTopologyEntity> getLldpLinkTopologyEntities();
+    List<OspfLinkTopologyEntity> getOspfLinkTopologyEntities();
 }
