--- opennms-dao-api/src/main/java/org/opennms/netmgt/dao/api/TopologyEntityCache.java
@@ -1,58 +1,0 @@
-/*******************************************************************************
- * This file is part of OpenNMS(R).
- *
- * Copyright (C) 2018 The OpenNMS Group, Inc.
- * OpenNMS(R) is Copyright (C) 1999-2018 The OpenNMS Group, Inc.
- *
- * OpenNMS(R) is a registered trademark of The OpenNMS Group, Inc.
- *
- * OpenNMS(R) is free software: you can redistribute it and/or modify
- * it under the terms of the GNU Affero General Public License as published
- * by the Free Software Foundation, either version 3 of the License,
- * or (at your option) any later version.
- *
- * OpenNMS(R) is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU Affero General Public License for more details.
- *
- * You should have received a copy of the GNU Affero General Public License
- * along with OpenNMS(R).  If not, see:
- *      http://www.gnu.org/licenses/
- *
- * For more information contact:
- *     OpenNMS(R) Licensing <license@opennms.org>
- *     http://www.opennms.org/
- *     http://www.opennms.com/
- *******************************************************************************/
-
-package org.opennms.netmgt.dao.api;
-
-import java.util.List;
-
-import org.opennms.netmgt.model.CdpLinkTopologyEntity;
-import org.opennms.netmgt.model.IsIsLinkTopologyEntity;
-import org.opennms.netmgt.model.LldpLinkTopologyEntity;
-import org.opennms.netmgt.model.NodeTopologyEntity;
-import org.opennms.netmgt.model.OspfLinkTopologyEntity;
-
-/**
- * Caches TopologyEmtities. This is a cache wrapper around @{@link TopologyEntityDao}. See there for an explanation of
- * TopologyEntrities.
- * We use the cache to improve the displaying speed of topologies.
- */
-public interface TopologyEntityCache {
-
-    List<NodeTopologyEntity> getNodeTopolgyEntities();
-
-    List<CdpLinkTopologyEntity> getCdpLinkTopologyEntities();
-
-    List<OspfLinkTopologyEntity> getOspfLinkTopologyEntities();
-
-    List<IsIsLinkTopologyEntity> getIsIsLinkTopologyEntities();
-
-    List<LldpLinkTopologyEntity> getLldpLinkTopologyEntities();
-
-    void refresh();
-
-}
